requirements.txt
fastapi==0.121.0
google-genai==1.49.0
pydantic==2.12.4
python-dotenv==1.2.1
requests==2.32.5
starlette==0.49.3
uvicorn==0.38.0
openai>=1.0.0
cohere>=4.0.0